requirements.txt
beautifulsoup4==4.6.0
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.7
django-ace==1.0.3
django-admin-bootstrapped==2.5.7
django-admin-tools==0.8.1
django-allauth==0.34.0
django-crispy-forms==1.7.0
django-grappelli==2.10.1
django-haystack==2.6.1
django-jet==1.0.6
django-pygments==0.1
django-suit==0.2.25
django-webtest==1.9.2
djangocodemirror==1.1.0
feedparser==5.2.1
google-api-python-client==1.4.1
httplib2==0.10.3
idna==2.6
Markdown==2.6.9
oauth2client==4.1.2
oauthlib==2.0.6
Pillow==4.3.0
psycopg2==2.6.2
pyasn1==0.3.7
pyasn1-modules==0.1.5
Pygments==2.2.0
PyJWT==1.5.3
pysolr==3.6.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
six==1.11.0
social-auth-app-django==2.0.0
social-auth-core==1.5.0
sorl-thumbnail==12.3
uritemplate==3.0.0
urllib3==1.22
waitress==1.1.0
WebOb==1.7.3
WebTest==2.0.29
